WIP: Monorepo to separate into smaller packages for easier maintenance
