+++ lerna.json
@@ -1,0 +1,9 @@
+{
+  "lerna": "3.4.0",
+  "packages": [
+    "packages/*"
+  ],
+  "npmClient": "yarn",
+  "useWorkspaces": true,
+  "version": "1.0.0-alpha1"
+}
