+++ packages/core/package.json
@@ -1,9 +1,30 @@
 {
-  "private": true,
-  "name": "tg-i18n-packages",
-  "workspaces": [
-    "packages/*"
+  "name": "tg-i18n",
+  "version": "1.0.0-alpha1",
+  "description": "Tool to generate translation files and provide them to React apps.",
+  "main": "dist/index.js",
+  "module": "dist/index.esm.js",
+  "typings": "dist/index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/thorgate/tg-i18n.git"
+  },
+  "keywords": [
+    "tg-i18n",
+    "react",
+    "react-dom",
+    "i18n",
+    "translations",
+    "gettext",
+    "higher order component",
+    "hoc"
   ],
+  "author": "Thorgate <hi@thorgate.eu>",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/thorgate/tg-i18n/issues"
+  },
+  "homepage": "https://github.com/thorgate/tg-i18n",
   "devDependencies": {
     "@types/hoist-non-react-statics": "3.0.1",
     "@types/jest": "23.3.2",
@@ -35,7 +56,6 @@
     "ts-jest": "23.10.1",
     "tsc-watch": "1.0.27",
     "tslint": "5.11.0",
-    "tslint-config-standard": "8.0.1",
     "tslint-react": "3.6.0",
     "typescript": "3.0.3"
   },
@@ -45,17 +65,14 @@
   },
   "dependencies": {
     "create-react-context": "0.2.3",
-    "date-fns": "1.29.0",
-    "gettext-extractor": "3.3.2",
-    "gettext-parser": "2.0.0",
     "hoist-non-react-statics": "3.0.1",
     "jed": "1.1.1",
     "prop-types": "15.6.2",
-    "shelljs": "0.8.2",
-    "tslib": "1.9.3",
-    "warning": "4.0.2",
-    "yargs": "12.0.2"
+    "warning": "4.0.2"
   },
+  "files": [
+    "dist/"
+  ],
   "jest": {
     "collectCoverageFrom": [
       "src/**/*.{ts,tsx}"
@@ -77,12 +94,21 @@
     ]
   },
   "scripts": {
-    "build": "lerna exec yarn build",
-    "clean": "lerna exec yarn clean",
-    "lint": "lerna exec yarn lint",
-    "test": "lerna exec yarn test",
-    "test:coverage": "lerna exec yarn test:coverage",
-    "check-packages": "yarn clean && yarn build && yarn test && yarn lint",
-    "publish": "lerna publish --skip-npm --force-publish '*'"
+    "clean": "$(npm bin)/rimraf dist",
+    "lint": "$(npm bin)/tslint -p .",
+    "watch": "rollup -cw",
+    "test": "$(npm bin)/jest",
+    "coverage": "npm run test -- --coverage",
+    "coveralls": "npm run coverage && cat ./coverage/lcov.info | $(npm bin)/coveralls",
+    "prebuild": "npm run clean -s && npm run lint && tsc -p tsconfig.base.json",
+    "build": "rollup -c",
+    "prepublish": "npm run build",
+    "docs": "$(npm bin)/jsdoc2md -t _docs/README.hbs src/*.js",
+    "prepare": "npm run build -s",
+    "prepublishOnly": "npm run build -s",
+    "patch": "npm version patch && npm publish",
+    "minor": "npm version minor && npm publish",
+    "major": "npm version major && npm publish",
+    "postpublish": "git push origin master --follow-tags"
   }
 }
