+++ src/utils.ts
@@ -1,3 +1,32 @@
+export const HeaderKeyOrder = [
+    'content-type',
+    'project-id-version',
+    'report-msgid-bugs-to',
+    'pot-creation-date',
+    'po-revision-date',
+    'last-translator',
+    'language-team',
+    'language',
+    'mime-version',
+    'content-transfer-encoding',
+    'plural-forms',
+];
+
+
 export function hasKey<O>(obj: O, key: string): key is Extract<keyof O, string> {
     return key in obj
+}
+
+
+export function makeArray<T>(value: T, length: number): T[] {
+    const array = [];
+    let i = 0;
+
+    while (i < length) {
+        array.push(value);
+
+        i += 1;
+    }
+
+    return array;
 }
