+++ tsconfig.base.json
@@ -2,13 +2,13 @@
   "compilerOptions": {
     "allowSyntheticDefaultImports": true,
     "declaration": true,
-    "declarationDir": "dist",
+    "declarationDir": "./lib",
     "emitDecoratorMetadata": false,
     "experimentalDecorators": false,
     "esModuleInterop": true,
     "importHelpers": true,
     "jsx": "react",
-    "lib": ["dom", "es2017"],
+    "lib": ["es2015", "es2016", "es2017", "dom"],
     "module": "es6",
     "moduleResolution": "node",
     "noFallthroughCasesInSwitch": true,
@@ -17,7 +17,7 @@
     "noImplicitThis": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
-    "outDir": "compiled",
+    "outDir": "./lib",
     "pretty": true,
     "removeComments": true,
     "sourceMap": true,
@@ -26,6 +26,6 @@
     "stripInternal": true,
     "target": "es2017"
   },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist"]
+  "include": ["src/**/*"],
+  "exclude": ["node_modules"]
 }
