+++ tslint.json
@@ -1,5 +1,7 @@
 {
   "extends": [
+    "tslint:recommended",
+    "tslint-config-standard",
     "tslint-react"
   ],
   "linterOptions": {
